Add a form; Refactoring app structure
