+++ functions/controllers/index.js
@@ -1,0 +1,6 @@
+const express = require('express')
+  , router = express.Router()
+
+router.use('/forms', require('./forms'))
+
+module.exports = router
