--- functions/routes/index.js
@@ -1,3 +1,0 @@
-module.exports = (app) => {
-  app.use('/forms', require('./forms'));
-};
